Move game components out of main
