+++ src/render.rs
@@ -1,7 +1,8 @@
 use macroquad::prelude::*;
 use shipyard::{Get, IntoIter, UniqueView, View, World};
 
-use crate::{game::{Game, BRUTE_SPAWN_HEALTH, PLAYER_RAY_LINGER, PLAYER_RAY_WIDTH}, physics::{BeamTag, BodyKind, ColliderTy, OneSensorTag, BodyTag}, BoxTag, BruteTag, BulletTag, EnemyState, Health, PlayerDamageState, PlayerGunState, PlayerScore, PlayerTag, RayTag, TileStorage, TileType, Transform};
+use crate::{physics::{BeamTag, BodyKind, ColliderTy, OneSensorTag, BodyTag}};
+use crate::game::*;
 // use macroquad_particles::{self as particles, BlendMode, ColorCurve, EmitterConfig};
 
 pub const WALL_COLOR: Color = Color::from_rgba(51, 51, 84, 255);
