+++ src/util.rs
@@ -1,8 +1,8 @@
 
 #[macro_export]
 macro_rules! inline_tilemap {
-    (@tile w) => { crate::TileType::Wall };
-    (@tile g) => { crate::TileType::Ground };
+    (@tile w) => { crate::game::TileType::Wall };
+    (@tile g) => { crate::game::TileType::Ground };
     (@tile $i:ident) => { $i };
     ($($tile:ident),+) => {
         vec![
